feat: add Twitter thread reader script
Add a new userscript to read Twitter threads in one single page. The script sends a GET request to https://twitter-thread.com/api/unroll-thread?id=<threadId> to get all the threads of the current post, and redirects the page to https://twitter-thread.com/t/<threadId> to display the threads in a new page.

The script is added as a new userscript called "Twitter thread reader", and registered with Greasemonkey for the Twitter post pages.

Issue: N/A